+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## [v1.6.2](https://github.com/codeducks/xbot/releases/tag/v1.6.2) (unreleased)
 
-* added guild-specific prefixes.
+* added config command (old bot command is now info)
 
 ## [v1.6.1](https://github.com/codeducks/xbot/releases/tag/v1.6.1) (26/1/2021)